% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dendro.R
\name{geom_dendro}
\alias{geom_dendro}
\title{create a dendrogram as several geom_path layers ready to be added to a ggplot object}
\usage{
geom_dendro(
  data,
  xlim = NULL,
  ylim = NULL,
  pointing = "updown",
  clustmethod = "complete",
  axis.labels = TRUE,
  ...
)
}
\arguments{
\item{data}{data.frame, columns will be clustered based on their rows.}

\item{xlim, }{vector with 2 numbers, on the x axis the dendrogram will beginn at the first number and end at the second.}

\item{ylim, }{vector with 2 numbers, on the y axis the dendrogram will beginn at the first number and end at the second.}

\item{pointing, }{string, either "side" or "updown" (default) to indicate where the dendrogram should point.}

\item{clustmethod}{string, method to be used by the hclust function (check ?hclust for available methods).}

\item{axis.labels, }{boolean, whether or not the axis should show the column names of data. This adds another layer to define the axis labels.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
a list of several ggplot2 layer objects (geom_path for the dendrogram) that can directly be added to a ggplot2 object
}
\description{
Imports:
ggplot2
}
\details{
the function uses geom_path for the dendrogram, so ... takes all arguements that geom_path would also take, such as color, size, etc.
}
\examples{
library(ggdendroplot)
library(ggplot2)

df <- matrix(rnorm(128), ncol = 8)
colnames(df) <- paste0("a",seq(ncol(df)))

ggplot() + geom_dendro(data=df)
ggplot() + geom_dendro(data=df, pointing="side")
ggplot() + geom_dendro(data=df, xlim=c(3,0))
ggplot() + geom_dendro(data=df, ylim=c(3,0))
ggplot() + geom_dendro(data=df, size=2, color="blue", linetype="dashed")
ggplot() + geom_dendro(data=df, size=4, lineend="round")

}
