% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmReady.R
\name{hmReady}
\alias{hmReady}
\title{creates a heatmap-ready data.frame}
\usage{
hmReady(df, rowclus = NULL, colclus = NULL)
}
\arguments{
\item{df}{data.frame}

\item{rowclus}{hclust object, use only if rows should be clustered}

\item{colclus}{hclust object, use only if columns should be clustered}
}
\value{
a data.frame with x, y and value column, ready to be plotted by geom_tile.
}
\description{
Imports:
reshape2
}
\examples{
library(ggdendroplot)

df <- as.data.frame(matrix(rnorm(99), ncol = 9))
rownames(df) <- paste0("trait", seq(nrow(df)))

clust <- hclust(dist(df))
hm <- hmReady(df, rowclus=clust)

ggplot(hm, aes(x, y, fill=value)) + geom_tile()

}
