% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hm.R
\name{theme_hm}
\alias{theme_hm}
\title{applies a heatmap-friendly theme to ggplot}
\usage{
theme_hm()
}
\value{
a ggplot layer that can be added to a ggplot object
}
\description{
Imports:
ggplot2
}
\examples{
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_hm()

}
